refactor: change aside componetnt logic from context to store
