+++ src/componets/pages/Layout/Layout.tsx
@@ -1,6 +1,5 @@
 import { Outlet } from 'react-router-dom';
 
-import { DataProvider } from '../../App/DataProvider/DataProvider';
 import Header from '../../widgets/Header/Header';
 import MainSection from '../../widgets/MainSection/MainSection';
 import StarBackground from '../../shared/UI/StarBackground/StarBackground';
@@ -9,14 +8,14 @@ import styles from './Layout.module.css';
 
 const Layout = () => {
   return (
-    <DataProvider>
+    <>
       <Header />
       <main className={styles.main}>
         <StarBackground />
         <MainSection />
         <Outlet />
       </main>
-    </DataProvider>
+    </>
   );
 };
