--- src/componets/shared/api/index.ts
@@ -1,3 +1,0 @@
-import getSingleCharacter from './getSingleCharacter';
-
-export default { getSingleCharacter };
